### 2021-07-10 #3470 @Kevin-Mattheus-Moerman
@savage13 thanks for this submission.

- [ ] Check `...released as under an...`, this should probably read `...released under an...`
- [ ] Please work to expand your statement of need to specifically state what scientific research this work enables. Also expand the paper in general with a more detailed discussion of related scientific research and or scientific research that already used this work. Use citations to the literature in this expanded text. See also our guidelines: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain.
