+++ paper/paper.bib
@@ -1,6 +1,50 @@
 
+@article{al2008improving,
+	author = {Al-Amri, Abdullah M and Rodgers, Arthur J and Al-Khalifah, Tariq A},
+	date-added = {2021-07-13 10:17:12 -0400},
+	date-modified = {2021-07-13 11:31:44 -0400},
+	doi = {10.1007/s12517-008-0001-5},
+	journal = {Arabian Journal of Geosciences},
+	number = {1},
+	pages = {1--15},
+	publisher = {Springer},
+	title = {{Improving the level of seismic hazard parameters in Saudi Arabia using earthquake location}},
+	volume = {1},
+	year = {2008},
+}
+
+@article{colombelli2013application,
+	author = {Colombelli, Simona and Allen, Richard M and Zollo, Aldo},
+	date-added = {2021-07-13 10:18:30 -0400},
+	date-modified = {2021-07-13 11:31:51 -0400},
+	doi = {doi:10.1029/2011GL047947},
+	journal = {Journal of Geophysical Research: Solid Earth},
+	number = {7},
+	pages = {3448--3461},
+	publisher = {Wiley Online Library},
+	title = {{Application of real-time GPS to earthquake early warning in subduction and strike-slip environments}},
+	volume = {118},
+	year = {2013},
+}
+
+@article{cramer2014update,
+	author = {Cramer, Chris H and Van Arsdale, Roy B and Dhar, Mahesh S and Pryne, Daniel and Paul, Justin},
+	date-added = {2021-07-13 11:36:42 -0400},
+	date-modified = {2021-07-13 11:38:01 -0400},
+	doi = {10.1785/0220160028},
+	journal = {Seismological Research Letters},
+	number = {5},
+	pages = {986--996},
+	publisher = {Seismological Society of America},
+	title = {Update of Urban Seismic-Hazard Maps for Memphis and Shelby County, Tennessee: Geology and VS Observations},
+	volume = {85},
+	year = {2014},
+}
+
 @techreport{goldstein1995status,
 	author = {Goldstein, P. and Minner, L.},
+	date-added = {2014-06-12 10:39:18 -0400},
+	date-modified = {2014-06-12 10:46:45 -0400},
 	doi = {10.2172/110248},
 	institution = {Lawrence Livermore National Lab., CA (United States)},
 	month = {08},
@@ -11,6 +55,8 @@
 
 @article{goldstein2003sac2000,
 	author = {Goldstein, P. and Dodge, D. and Firpo, M. and Minner, L.},
+	date-added = {2014-06-12 09:49:06 -0400},
+	date-modified = {2014-06-12 09:49:57 -0400},
 	journal = {The IASPEI International Handbook of Earthquake and Engineering Seismology},
 	pages = {1613--1620},
 	publisher = {Academic Press, London},
@@ -21,6 +67,8 @@
 
 @article{goldstein2005sac,
 	author = {P. Goldstein and A. Snoke},
+	date-added = {2014-06-12 09:38:38 -0400},
+	date-modified = {2014-06-12 09:40:56 -0400},
 	journal = {Incorporated Institutions For Seismology Data Management Center Electronic Newsletter},
 	number = {1},
 	title = {{SAC Availability for the IRIS Community}},
@@ -28,3 +76,61 @@
 	volume = {7},
 	year = {2005},
 }
+
+@article{miao2007empirical,
+	author = {Miao, Qingwen and Langston, Charles A},
+	date-added = {2021-07-13 10:16:34 -0400},
+	date-modified = {2021-07-13 10:30:57 -0400},
+	doi = {10.1785/0120110042},
+	journal = {Bulletin of the Seismological Society of America},
+	number = {6},
+	pages = {2137--2151},
[…]
