### 2021-07-13 #3470 @savage13
Fixed the following two items:

* [x]  Check `...released as under an...`, this should probably read `...released under an...`

* [x]  Please work to expand your statement of need to specifically state what scientific research this work enables. Also expand the paper in general with a more detailed discussion of related scientific research and or scientific research that already used this work. Use citations to the literature in this expanded text. See also our guidelines: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain.

A paragraph was added to the beginning of the 'Statement of Need' with example research applications of this library.
