### 2021-07-19 #3470 @danielskatz
👋  @rkurchin - would you be able to edit this submission?  (Note that @kbarnhart has [volunteered to help find reviewers](https://github.com/openjournals/joss-reviews/issues/3470#issuecomment-878339073))
