### 2021-07-19 #3470 @rkurchin
I don't know C or anything about seismic analysis, so I doubt I'm a good fit either...
