### 2021-08-04 #3470 @danielskatz
I'll take this on, with help accepted from @kbarnhart on reviewers
