### 2021-08-10 #3470 @kbarnhart
Some other recommendations @danielskatz.
- [Robert Herrmann @ SLU]( https://www.slu.edu/arts-and-sciences/earth-atmospheric-sciences/faculty/herrmann-robert.php)
- [Charles Amon @ Penn State]( http://eqseis.geosc.psu.edu/cammon/)
- [Chad Trabant at IRIS](https://www.iris.edu/hq/staff/employee/trabant) chad-iris
- jkmacc-LANL (listed above already) and mbegnaud (contributors to [pySAC](https://github.com/LANL-Seismoacoustics/pysac))
