### 2021-08-10 #3470 @danielskatz
👋 @chad-iris and @mbegnaud - Would one or both of you be able to review this submission to JOSS?
