### 2021-08-13 #3470 @danielskatz
👋 @cja12 - would you (or perhaps a member of your group) be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
