### 2021-08-16 #3470 @cja12
This is a very busy time of the year with classes starting. I can try to review this package, but cannot do anything extensive.

> On Aug 16, 2021, at 8:42 AM, Daniel S. Katz ***@***.***> wrote:
>
>
> 👋 @krischer, @jkmacc-LANL, @mbegnaud, @cja12 - I'm just checking with you again to see if one of you could be the second reviewer for this JOSS submission
>
