### 2021-08-16 #3619 @danielskatz
👋 @savage13 - can you work on the possibly missing DOI that whedon suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
