### 2021-08-16 #3619 @danielskatz
@cja12 - I see you are trying to use the automated GitHub feature on the checklist - these issues you open are intentionally being closed by whedon.  Please note the instructions at the top of your checklist:

>✨ Important: Please do not use the _Convert to issue_ functionality when working through this checklist, instead, please open any new issues associated with your review <a target="_blank" href="https://github.com/savage13/sacio">in the software repository associated with the submission</a>. ✨

and when you open an issue there, please mention `openjournals/joss-reviews#3619` so that a link is created to this review issue and we can see if it is open or closed.
