### 2021-08-16 #3470 @cja12
Hi,

I am sorry. Please remove me from this review process. This is a mess and now I am being spammed continuously with emails from your system. I don’t have time to navigate such a convoluted review process.

Chuck Ammon

> On Aug 16, 2021, at 8:42 AM, Daniel S. Katz ***@***.***> wrote:
>
>
> 👋 @krischer, @jkmacc-LANL, @mbegnaud, @cja12 - I'm just checking with you again to see if one of you could be the second reviewer for this JOSS submission
>
