### 2021-08-16 #3619 @danielskatz
@cja12 - Did you see the instructions in https://github.com/openjournals/joss-reviews/issues/3619#issuecomment-899602160?  Right now, you are getting notifications for all JOSS reviews, and those instructions tell you how to turn this off, so you will only get notifications for this review.  I hope that might solve your problem and let you keep working on this review, particularly since you already have started.
