+++ paper/paper.bib
@@ -109,7 +109,7 @@
 	author = {Tape, C. and Liu, Q. and Maggi, A. and Tromp, J.},
 	date-added = {2011-03-16 16:16:15 -0400},
 	date-modified = {2021-07-13 11:33:39 -0400},
-	dot = {10.1111/j.1365-246X.2009.04429.x},
+	doi = {10.1111/j.1365-246X.2009.04429.x},
 	issn = {1365-246X},
 	journal = {Geophys. Jour. Int.},
 	number = {1},
