### 2021-09-08 #3619 @danielskatz
👋 @chad-iris, @cja12, @jkmacc-LANL - we're now about 3 weeks into the process.  How is it going?  I see some progress by @cja12, but no checked items from @chad-iris or @jkmacc-LANL
