### 2021-09-13 savage13/sacio#1 @jkmacc-LANL
Installation instructions missing
Hello!  While there is an [Downloading and installing](https://github.com/savage13/sacio#downloading-and-installing) section in your project's README, there appear to be no instructions to download or install the software.  Many users will certainly be familiar with Makefiles and compiling C code, but some may not.  In your README, could you please provide a few sentences of instruction or CLI commands, as well as any pre-installation requirements?

Thank you for your efforts on this library!

EDIT: cross-link to openjournals/joss-reviews#3619
