### 2021-09-13 9cd687b
Added documentation for compiling and installation
Fixes to #1 with instructions for downloading, compiling and installation.
