### 2021-09-13 b9d23fc
Add language about dependencies
Link to #1
