### 2021-09-29 #3619 @danielskatz
👋 @chad-iris, @cja12, @jkmacc-LANL - we're now about 6 weeks into the process. How is it going? I see some progress by @cja12 and @jkmacc-LANL, but no checked items from @chad-iris.

Are there things that you need from the author or me?
