### 2021-09-29 #3619 @cja12
Hi,

I informed you weeks ago that I was through using your system. I do not have any plans to continue this process. I found the interface too confusing and annoying - especially when it automated sending me dozens of irrelevant messages to me. For example, other than to me, I have no idea where this response is going. Ridiculous.

> On Sep 29, 2021, at 1:55 PM, Daniel S. Katz ***@***.***> wrote:
>
>
> 👋 @chad-iris, @cja12, @jkmacc-LANL - we're now about 6 weeks into the process. How is it going? I see some progress by @cja12 and @jkmacc-LANL, but no checked items from @chad-iris.
>
> Are there things that you need from the author or me?
>
