### 2021-10-15 savage13/sacio#2 @chad-iris
Paper typos and suggestions
- "seismoloigical" -> seismological

- "restrited" -> restricted, may also consider writing "export-restricted"

- "used by seismologist globally" -> "seismologists" (plural)

- Inconsistent use of "meta-data" and "meta data", personally I prefer "metadata"

- "2 Clause" -> "2-Clause" or "2-clause"
