### 2021-10-16 savage13/sacio#4 @chad-iris
Context in documentation/README
The README in the repo and same documentation at https://savage13.github.io/sacio/html/index.html does not provide much context for what "sac files" are.

A few sentences from the paper in the Overview section would help anyone stumbling on this repository understand that sac files are a seismic data format in this context.
