### 2021-10-16 savage13/sacio#5 @chad-iris
Summary of paper, clarification of data format only scope
First sentence in the Summary of the paper describes the sac processing software and file format.  Then some talk of format and justification.  Finally:

| Here we implement an improved and open source version with backwards compatibility.

Explicitly stating that the implementation is for the format only may add clarity.  Suggestion:

| Here we implement an improved and open source version of the sac format with backwards compatibility.
