### 2021-11-02 #3619 @mbegnaud
I will try to get it to you within a week

________________________________
Dr. Michael Begnaud
LANL Project Manager, Ground-based Nuclear Detonation Detection
LANL Venture Lead, Dynamic Networks, Low-yield Nuclear Monitoring
Los Alamos National Laboratory ( http://www.lanl.gov<http://www.lanl.gov/> )
P.O. Box 1663  MS F665 (Mailing),  Bikini Atoll Rd., SM 30 (Shipping)
Los Alamos, NM  87545
(505) 667-7620, ***@***.******@***.***>
***@***.***>
***@***.***>
