### 2021-11-08 d880efe
Fix typos; resolve #2
