### 2021-11-08 0c7b96c
Added section in README to describe SAC files; resolve #4
