### 2021-11-08 cefca2e
Added clarify to last sentence of the summary; resolve #5
