### 2021-11-08 81e5359
Updated example in README.md; takes a file a a command line argument; resolve #3
