+++ README.md
@@ -29,53 +29,75 @@ Examples
 
 ### Read, Change and Write a sac file
 
-```c
-#include <stdio.h>
+Compile and run the example below using:
 
+   gcc -I. -o example example.c libsacio_bsd.a
+   ./example t/imp.sac
+
+The file `t/imp.sac` is an impulse function and distributed with `sacio`.
+
+```c
+
+#include <stdio.h>
 #include <sacio.h>
 
-// Variable Declaration
-int nerr = 0;
-char org[128] = {0};
-char id[64] = {0};
-sac *s = NULL;
+int
+main(int argc, char *argv[]) {
+    // Variable Declaration
+    int i;
+    int nerr = 0;
+    char org[128] = {0};
+    char id[64] = {0};
+    double delta;
+    sac *s = NULL;
 
-// Read a SAc file, could be evenly or unevenly spaced
-s = sac_read("file.sac", &nerr);
-if(nerr != 0) {
-    printf("Error: %d reading file\n", nerr);
-    return;
+    if(argc < 2) {
+        printf("Usage: example file.sac\n");
+        return -1;
+    }
+
+    // Read a SAc file, could be evenly or unevenly spaced
+    s = sac_read(argv[1], &nerr);
+    if(nerr != 0) {
+        printf("Error: %d reading file\n", nerr);
+        return -1;
+    }
+
+    // Get a floating point value: delta
+    sac_get_float(s, SAC_DELTA, &delta);
+
+    // Print out some header values
+    printf("delta: %f\n", delta);
+    printf("npts:  %d\n", s->h->npts);
+    printf("comps: %d\n", sac_comps(s));
+
+    // Print out the first 5 data values
+    for(i = 0; i < 5; i++) {
+        printf("y[%4d]  %f\n", i, s->y[i]);
+    }
+
+    // Set some header character string values
+    sac_set_string(s, SAC_NET, "IU");
+    sac_set_string(s, SAC_STA, "BORG");
+    sac_set_string(s, SAC_LOC, "00");
+    sac_set_string(s, SAC_CHA, "BHZ");
+
+    // Set the origin time to 1994/160 00:33:16.123
+    sac_set_time(s, timespec64_from_yjhmsf(1994, 160, 0, 33, 16, 123));
+
+    // Format the origin time in absolute time
+    sac_fmt(org, sizeof(org), "%TO", s);
+    printf("Origin Time: %s [Absolute]\n", org);
+
+    // Output the Station id: NET.STA.LOC.CHA
+    sac_fmt(id, sizeof(id), "%Z", s);
+    printf("Station ID: %s\n", id);
+
+    // Write out the modified sac file
+    sac_write(s, "output.sac", &nerr);
+
+    return 0;
 }
-
-// Print out some header values
-printf("delta: %f\n", s->h->delta);
-printf("npts:  %d\n", s->h->npts);
-printf("comps: %d\n", sac_comps(s));
-
-// Print out the first 5 data values
-for(i = 0; i < 5; i++) {
-    printf("y[%4d]  %f\n", i, s->y[i]);
-}
-
-// Set some header character string values
-sac_set_string(s, SAC_NET, "IU");
-sac_set_string(s, SAC_STA, "BORG");
-sac_set_string(s, SAC_LOC, "00");
-sac_set_string(s, SAC_CHA, "BHZ");
-
-// Set the origin time to 1994/160 00:33:16.123
-sac_set_time(s, timespec64_from_yjhmsf(1994, 160, 0, 33, 16, 123));
-
-// Format the origin time in absolute time
-sac_fmt(org, sizeof(ostr), "%TO", s);
[…]
