### 2021-11-18 #3619 @mbegnaud
The README.md file seems to have the "Examples" part out of order with the Installation part. Can't run the Example until it is installed correct? If I run the configure command with the "--prefix" to another location, need to be sure the user has that path in their environment. That is not clear in the instructions.
