### 2021-11-18 #3619 @mbegnaud
Also in the README.md, seems cleaner to have 2 separate lines for compiling the example and then running. I'm on a Mac. I had to change so the "-I" had the include directory from where I installed the files. Maybe the example.c file should be created in the same directory as where the source code lies?? What should the "example" program output so the user can compare??
