### 2021-11-18 #3619 @mbegnaud
I don't see documentation for read-world examples except the one. If there is a document or file somewhere that describes this, it needs to be included as a link or similar in the README.md file.
