### 2021-11-18 savage13/sacio#6 @mbegnaud
README.md and Example file
The README.md file seems to have the "Examples" part out of order with the Installation part. Can't run the Example until it is installed correct? If I run the configure command with the "--prefix" to another location, need to be sure the user has that path in their environment. That is not clear in the instructions.

seems cleaner to have 2 separate lines for compiling the example and then running. I'm on a Mac. I had to change so the "-I" had the include directory from where I installed the files. Maybe the example.c file should be created in the same directory as where the source code lies?? What should the "example" program output so the user can compare??

openjournals/joss-reviews#3619
