### 2021-11-18 savage13/sacio#7 @mbegnaud
Don't see Real-world examples of use except for 1
I don't see documentation for real-world examples except the one. If there is a document or file somewhere that describes this, it needs to be included as a link or similar in the README.md file. If there isn't, there needs to include some distinct examples of use.

openjournals/joss-reviews#3619
