### 2021-11-18 savage13/sacio#8 @mbegnaud
Not clear on community guidelines
In reviewer instructions, it is not clear that this step is made:

Should be in REAME.md or other file. If it is there, please highlight.

openjournals/joss-reviews#3619
