### 2021-11-18 savage13/sacio#9 @mbegnaud
Link or show References for paper.md file
A link to the references or the actual references need to be shown for the paper.md file.

openjournals/joss-reviews#3619
