### 2021-11-22 savage13/sacio#9 @danielskatz
I don't think I understand this.  Can you be more specific?

I wonder if you are looking at the difference between [paper.md](https://github.com/savage13/sacio/blob/master/paper/paper.md), which is source that doesn't have a references section, and the built paper (currently https://github.com/openjournals/joss-papers/blob/joss.03619/joss.03619/10.21105.joss.03619.pdf), which does have the full references.
