### 2021-11-23 80ebd3e
Add references to example code, resolve #7
