### 2021-11-23 0d5ae20
Add  info about compiling the example, a reference to compiling the library, and if installation is needed, resolve #6
