### 2021-11-23 41adf16
Updated community guidelines highlighing bugs, contributions and support. resolve #8
