### 2021-11-23 #3619 @savage13
Updated and fixed issues opened by mbegnaud.  Sorry for the delay.
