### 2021-11-23 #3619 @danielskatz
@savage13 - Next I'll proofread the paper, provide any comments, and provide instruction on archiving the software, which is what we need to proceed to acceptance.
